removed gecko, added owl files to mappings
- removed references to the template-based gecko build from the Makefile, and replaced by a simple download of GECKO by its purl.
- Makefile now copies owl files into a new directory `data_dictionaries`
- mapping tsvs are updated
- odk.sh added with respective section in README.md that allows to run the pipeline in the odk conainer entirely
- added a missing property to properties.tsv

Fixed a bug in generate_cohort_json, but not sure whether the bug was actually indicative of some more deeply routed problem with one of the datasets. The scenario is this: a target came in, that had a key which was already associated - but with an empty list -> this made it fail in the recursion. The two lines now make sure that if the value stored with the key is None or [], it gets replaced instead with an empty dict.

There is still a bug with the data.json goal, but @beckyjackson agreed to fix it .

```
java -jar build/robot.jar --prefixes src/prefixes.json export \
--input build/gecko.owl \
--header "ID|LABEL|definition|question description|value|see also|subclasses" \
--sort "LABEL" \
--export build/browser/gecko-data.json
INVALID COLUMN ERROR unable to find property for column header 'question description'
For details see: http://robot.obolibrary.org/export#invalid-column-error
Use the -vvv option to show the stack trace.
Use the --help option to see usage information.
Makefile:318: recipe for target 'build/browser/gecko-data.json' failed
make: *** [build/browser/gecko-data.json] Error 1
```